asx-news-reaction-app/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.1
APScheduler==3.11.0
httpx==0.28.1
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
pandas==2.2.3
yfinance==0.2.54
vaderSentiment==3.3.2
psycopg2-binary==2.9.10
